Add support for configurable baudrate through run-time parameters
In our platform, physical access to the MTi is challanging so we want to
configure the MTi through the driver parameters. A previous patch added
the ability for output configuration. This may lead to higher bandwidth
requirements, hence the need to configure the baudrate.

To ensure the device can be detect initialy, and after potential factory
resets, an additional baudrate parameter is added to the configuration
and scan process. To ensure the device can be detected.